refactor(common-json): relocate DirectBufferInputStreamEx to common-agrona
DirectBufferInputStreamEx is a generic buffer-to-InputStream bridge with no
JSON semantics; move it to io.aklivity.zilla.runtime.common.agrona.io,
mirroring upstream agrona's own org.agrona.io package, alongside its *Ex
buffer siblings rather than living in common-json.

JsonNode no longer needs the InputStream bridge at all - it now feeds the
buffer directly via JsonEx.createParser().wrap(buffer, offset, length).

McpServerFactory and all common-json call sites/tests update their import
to the new location; no module-info changes are needed beyond exporting the
new package from common-agrona, since binding-mcp already gets transitive
access via engine's existing requires transitive on common-agrona.

Fixes #1999
