+++ runtime/common-agrona/src/main/java/io/aklivity/zilla/runtime/common/agrona/io/DirectBufferInputStreamEx.java
@@ -1,18 +1,19 @@
 /*
- * Copyright 2021-2024 Aklivity Inc
+ * Copyright 2021-2024 Aklivity Inc.
  *
- * Licensed under the Aklivity Community License (the "License"); you may not use
- * this file except in compliance with the License.  You may obtain a copy of the
- * License at
+ * Aklivity licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
  *
- *   https://www.aklivity.io/aklivity-community-license/
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
  */
-package io.aklivity.zilla.runtime.common.json;
+package io.aklivity.zilla.runtime.common.agrona.io;
 
 import java.io.InputStream;
