+++ runtime/common-agrona/src/main/moditect/module-info.java
@@ -19,4 +19,5 @@ module io.aklivity.zilla.runtime.common.agrona
 
     exports io.aklivity.zilla.runtime.common.agrona.buffer;
     exports io.aklivity.zilla.runtime.common.agrona.concurrent;
+    exports io.aklivity.zilla.runtime.common.agrona.io;
 }
