+++ runtime/common-json/src/test/java/io/aklivity/zilla/runtime/common/json/JsonParserTest.java
@@ -44,6 +44,7 @@ import jakarta.json.stream.JsonParsingException;
 import org.junit.jupiter.api.Test;
 
 import io.aklivity.zilla.runtime.common.agrona.buffer.UnsafeBufferEx;
+import io.aklivity.zilla.runtime.common.agrona.io.DirectBufferInputStreamEx;
 
 class JsonParserTest
 {
