+++ runtime/common-json/src/test/java/io/aklivity/zilla/runtime/common/json/JsonSchemaDependentTest.java
@@ -23,6 +23,7 @@ import jakarta.json.stream.JsonParser;
 import org.junit.jupiter.api.Test;
 
 import io.aklivity.zilla.runtime.common.agrona.buffer.UnsafeBufferEx;
+import io.aklivity.zilla.runtime.common.agrona.io.DirectBufferInputStreamEx;
 
 class JsonSchemaDependentTest
 {
