+++ runtime/common-json/src/test/java/io/aklivity/zilla/runtime/common/json/JsonValidatorChainTest.java
@@ -29,6 +29,7 @@ import org.junit.jupiter.api.Test;
 
 import io.aklivity.zilla.runtime.common.agrona.buffer.MutableDirectBufferEx;
 import io.aklivity.zilla.runtime.common.agrona.buffer.UnsafeBufferEx;
+import io.aklivity.zilla.runtime.common.agrona.io.DirectBufferInputStreamEx;
 import io.aklivity.zilla.runtime.common.json.JsonPipeline.Status;
 
 class JsonValidatorChainTest
